requirements.txt
openai>=1.0.0
tqdm
transformers
torch
accelerate
google-generativeai
python-dotenv
matplotlib

